feat: try to parse a realistic VCD file
Something is off with the parsing for:

$var integer 1 : write_en $end
