+++ vcd/samples/BUILD.bazel
@@ -1,6 +1,6 @@
-package(default_visibility = [ "//vcd:__pkg__"])
+package(default_visibility = ["//vcd:__pkg__"])
 
 filegroup(
     name = "samples",
-    srcs = glob(["*.vcd"])
+    srcs = glob(["*.vcd"]),
 )
